report

+++ FACULTY_ASSOCDEAN/ui-formsPreview.php
@@ -1,5 +1,6 @@
 <?php
   session_start();
+  include 'db.php';
 
   if(!isset($_SESSION['id'])) {
     header("location: ../");
@@ -16,91 +17,54 @@
 ?>
 <!DOCTYPE html>
 <html lang="en">
-
-<?php
-// Database connection details
-$db_host = "localhost"; // Replace with your database host
-$db_username = "root"; // Replace with your database username
-$db_password = ""; // Replace with your database password
-$db_name = "unilink"; // Replace with your database name
-
-// Connect to the database
-$conn = new mysqli($db_host, $db_username, $db_password, $db_name);
-
-// Check connection
-if ($conn->connect_error) {
-    die("Connection failed: " . $conn->connect_error);
-}
-
-// SQL query to fetch data from the table
-$sql = "SELECT * FROM activityform";
-$result = $conn->query($sql);
-?>
-
 <head>
-
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <meta name="description" content="">
     <meta name="author" content="">
-
     <title>UniLink - Faculty/Associate Dean</title>
     <link rel="shortcut icon" type="image/png" href="../imgs/BSU.png" alt="Logo" />
-
     <!-- Custom fonts for this template-->
     <link href="vendor/fontawesome-free/css/all.min.css" rel="stylesheet" type="text/css">
     <link
         href="https://fonts.googleapis.com/css?family=Nunito:200,200i,300,300i,400,400i,600,600i,700,700i,800,800i,900,900i"
         rel="stylesheet">
      <link href="css/sb-admin-2.min.css" rel="stylesheet">
-     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.5.2/css/bootstrap.css">
-        
+     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.5.2/css/bootstrap.css">   
     <link href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" rel="stylesheet">
-
     <!-- JQUERY -->
     <script src="https://code.jquery.com/jquery-3.7.1.min.js" integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo=" crossorigin="anonymous"></script>
-
     <!-- DATATABLES -->
     <!-- <link rel="stylesheet" href="https://cdn.datatables.net/1.13.6/css/jquery.dataTables.min.css"> -->
     <script src="https://cdn.datatables.net/1.13.6/js/jquery.dataTables.min.js"></script>
     <link href="vendor/datatables/dataTables.bootstrap4.min.css" rel="stylesheet">
-
     <script>
         $(document).ready(function() {
             $('#activityTable').DataTable();
         });
     </script>
-
     <!-- Custom styles for this template-->
     <link href="css/sb-admin-2.min.css" rel="stylesheet">
-
 </head>
-
 <body id="page-top">
-
     <!-- Page Wrapper -->
     <div id="wrapper">
-
         <!-- Sidebar -->
         <ul class="navbar-nav bg-gradient-primary sidebar sidebar-dark accordion" id="accordionSidebar">
-
             <!-- Sidebar - Brand -->
             <a class="sidebar-brand d-flex align-items-center justify-content-center" href="index.php">
                 <img src="..\imgs\BSU.png" height="45" width="50">
                 <div class="sidebar-brand-text mx-3">UniLink</div>
             </a>
-
             <!-- Divider -->
             <hr class="sidebar-divider my-0">
-
             <!-- Nav Item - Dashboard -->
             <li class="nav-item">
                 <a class="nav-link" href="index.php">
                     <i class="fas fa-fw fa-tachometer-alt"></i>
                     <span>Dashboard</span></a>
             </li>
-
             <li class="nav-item active">
                 <a class="nav-link" href="ui-formsPreview.php">
                     <i class="fas fa-fw fa-calendar"></i>
@@ -113,31 +77,24 @@ $result = $conn->query($sql);
                 </a>
             </li>
             <hr class="sidebar-divider">
-
             <!-- Sidebar Toggler (Sidebar) -->
             <div class="text-center d-none d-md-inline">
                 <button class="rounded-circle border-0" id="sidebarToggle"></button>
             </div>
-
         </ul>
         <!-- End of Sidebar -->
-
         <!-- Content Wrapper -->
         <div id="content-wrapper" class="d-flex flex-column">
-
             <!-- Main Content -->
             <div id="content">
-
                 <!-- Topbar -->
                 <nav class="navbar navbar-expand navbar-light bg-white topbar mb-4 static-top shadow">
-
                     <!-- Sidebar Toggle (Topbar) -->
                     <button id="sidebarToggleTop" class="btn btn-link d-md-none rounded-circle mr-3">
                         <i class="fa fa-bars"></i>
                     </button>
                     <!-- Topbar Navbar -->
                     <ul class="navbar-nav ml-auto">
-
                         <!-- Dropdown - Messages -->
                         <div class="dropdown-menu dropdown-menu-right p-3 shadow animated--grow-in"
                              aria-labelledby="searchDropdown">
@@ -159,7 +116,14 @@ $result = $conn->query($sql);
                         <li class="nav-item dropdown no-arrow">
                             <a class="nav-link dropdown-toggle" href="#" id="userDropdown" role="button"
                                 data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                                <span class="mr-2 d-none d-lg-inline text-gray-600 small">Douglas McGee</span>
+                                <?php
+                                    $id = $_SESSION['id'];
+
+                                    $sql = "SELECT * FROM `users`";
+                                    $result = $conn->query($sql);
+                                    $row = $result->fetch_assoc();
+                                ?>
+                                <span class="mr-2 d-none d-lg-inline text-gray-600 small"><?=$row['first_name']." ".$row['last_name']?></span>
                                 <img class="img-profile rounded-circle"
                                     src="imgs/undraw_profile.svg">
                             </a>
@@ -177,12 +141,8 @@ $result = $conn->query($sql);
                                 </a>
                             </div>
                         </li>
-
                     </ul>
-
                 </nav>
-
-
                 <!-- End of Topbar -->
                 <div class="container-fluid">
                     <div class="d-flex justify-content-between align-items-center mb-4">
@@ -210,22 +170,10 @@ $result = $conn->query($sql);
                                 </thead>
                                <tbody>
                         <?php
-                        // Replace with your database connection details
-                        $servername = "localhost";
-                        $username = "root";
-                        $password = "";
-                        $dbname = "unilink";
-
-                        // Create a connection to the database
-                        $conn = new mysqli($servername, $username, $password, $dbname);
-
-                        // Check connection
-                        if ($conn->connect_error) {
-                            die("Connection failed: " . $conn->connect_error);
-                        }
-
                         // SQL query to retrieve data
-                        $sql = "SELECT * FROM activityform";
+                        $college = $_SESSION['college'];
+
+                        $sql = "SELECT * FROM `activityform` WHERE `college` = '$college'";
                         $result = $conn->query($sql);
 
                         if ($result->num_rows > 0) {
@@ -240,7 +188,7 @@ $result = $conn->query($sql);
 
                                 if ($start_date > $today) {
                                     $status = 'For Implementation';
-                                    $textColor = 'black'; // Set text color for 'For Implementation'
+                                    $textColor = '#d7d0d0'; // Set text color for 'For Implementation'
                                 } elseif ($start_date <= $today && $end_date > $today) {
                                     $status = 'Ongoing';
                                     $textColor = 'orange'; // Set text color for 'Ongoing'
@@ -253,14 +201,26 @@ $result = $conn->query($sql);
                                 echo "<td style='text-align: center;'>" . $row["activity_title"] . "</td>";
                                 echo "<td style='text-align: center;'>" . $row["partner"] . "</td>";
                                 echo "<td style='text-align: center;'>" . date("M. d, Y", $start_date) . " - " . date("M. d, Y", $end_date) . " </td>";
-                                echo '<td style="text-align: center;"><div style="background-color: #d7d0d0; color: ' . $textColor . '; padding: 5px; border-radius: 45px;">' . $status . '</div></td>';
-                                echo "<td style='text-align: center;'>
+                                echo '<td style="text-align: center;"><div style="background-color: '.$textColor.'; color: #000000; padding: 5px; border-radius: 45px;">' . $status . '</div></td>';
+                                if($status == "Implemented") {
+                                    echo "
+                                        <td style='text-align: center;'>
+                                            <a href='ui-formsEdit.php?id=" . $row["id"] . "'>
+                                                <span class='fas fa-edit text-secondary' title='Edit'></span>
+                                            </a>
+                                            <a href='pdf.php?id=". $row["id"]."' target='_blank' class='fas fa-file-download text-info' title='Request Form'></a>
+                                            <a href='report.php?id=".$row['id']."' target='_blank' class='fas fa-clipboard text-success' title='Narrative Report'></a>
+                                        </td>
+                                    ";
+                                } else {
+                                    echo "<td style='text-align: center;'>
                                         <a href='ui-formsEdit.php?id=" . $row["id"] . "'>
-                                            <span class='fas fa-edit text-success'></span>
+                                            <span class='fas fa-edit text-secondary'></span>
                                         </a>
                                         <a href='delete.php?id=" . $row["id"] . "'><span class='fas fa-trash text-danger'></span></a>
                                         <a href='pdf.php?id=". $row["id"]."' target='_blank' class='fas fa-file-download text-info'></a>
                                     </td>";
+                                }
                                 echo "</tr>";
                                 $i++;
                             }
@@ -270,32 +230,25 @@ $result = $conn->query($sql);
 
                         $conn->close();
                         ?>
-
-
-
-                </tbody>
-
+                        </tbody>
                             </table>
                         </div>
                     </div>
                     </div>
                 </div>
-
+            </div>
                 <!-- End of Main Content -->
-
                 <!-- Footer -->
                 <footer class="sticky-footer bg-white">
                     <div class="container my-auto">
                         <div class="copyright text-center my-auto">
-                            <span>Developed by Unilink 2023</span>
+                            <span>Developed by UniLink 2023</span>
                         </div>
                     </div>
                 </footer>
                 <!-- End of Footer -->
-
             </div>
             <!-- End of Content Wrapper -->
-
         <!-- End of Page Wrapper -->
         <!-- Scroll to Top Button-->
         <a class="scroll-to-top rounded" href="#page-top">
@@ -320,20 +273,15 @@ $result = $conn->query($sql);
                 </div>
             </div>
         </div>
-
         <!-- Bootstrap core JavaScript-->
         <script src="vendor/jquery/jquery.min.js"></script>
         <script src="vendor/bootstrap/js/bootstrap.bundle.min.js"></script>
-
         <!-- Core plugin JavaScript-->
         <script src="vendor/jquery-easing/jquery.easing.min.js"></script>
-
         <!-- Custom scripts for all pages-->
         <script src="js/sb-admin-2.min.js"></script>
-
         <!-- Page level plugins -->
         <script src="vendor/chart.js/Chart.min.js"></script>
-
         <!-- Page level custom scripts -->
         <script src="js/demo/chart-area-demo.js"></script>
         <script src="js/demo/chart-pie-demo.js"></script>
@@ -341,5 +289,4 @@ $result = $conn->query($sql);
     include 'footer.php';
 ?>
     </body>
-
 </html>
